Reorganized repository. Added setup file
