+++ .gitignore
@@ -1,3 +1,9 @@
 results/**
 kernels/**
 !kernels/sample_kernel.py
+
+# Ignore egg-info files
+*.egg-info
+
+# Ignore all __pycache__ folders in the project
+__pycache__/
